HIS-80: Fix infinite fetching bug
