+++ histree-frontend/src/components/comparison/ComparisonCard.tsx
@@ -5,18 +5,18 @@ import ComparisonFlow from './ComparisonFlow';
 import { ReactFlowProvider } from 'reactflow';
 
 export const ComparisonCard = () => {
-  return (
-    <div className="comparison-container">
-      <Card>
-        <CardHeader title="Relationship Search" />
-        <CardContent>
-          <div className="comparison-container-inner">
-            <ReactFlowProvider>
-              <ComparisonFlow />
-            </ReactFlowProvider>
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  );
+	return (
+		<div className="comparison-container">
+			<Card>
+				<CardHeader style={{ padding: '1em 0 0 1em', margin: 0 }} title="Relationship Search" />
+				<CardContent>
+					<div className="comparison-container-inner">
+						<ReactFlowProvider>
+							<ComparisonFlow />
+						</ReactFlowProvider>
+					</div>
+				</CardContent>
+			</Card>
+		</div>
+	);
 };
