+++ histree-frontend/src/components/comparison/ComparisonToggle.scss
@@ -1,9 +1,5 @@
 .comparison_toggle_container {
-	margin: 1em;
+	margin: 1.5em 0.5em 0 0;
 	width: 5%;
 	height: 5%;
-	position: absolute;
-	z-index: 98;
-	left: 1%;
-	top: 3%;
 }
