+++ histree-frontend/src/components/general/SearchBar.scss
@@ -1,9 +1,5 @@
 .search_container {
 	margin: 1em;
-	width: 35%;
+	width: 400px;
 	height: 3.5em;
-	position: absolute;
-	z-index: 99;
-	left: 5%;
-	top: 2%;
 }
