+++ histree-frontend/src/pages/TreePage.scss
@@ -11,3 +11,11 @@
 		padding: 2em;
 	}
 }
+
+.topleft-container {
+	position: absolute;
+	z-index: 99;
+	left: 5%;
+	top: 2%;
+	display: flex;
+}
